+++ CHANGELOG.md
@@ -20,6 +20,10 @@
   [@freiksenet](https://github.com/freiksenet) in [#1003](https://github.com/apollographql/graphql-tools/pull/1003)
 * Allow user-provided `buildSchema` options.  <br/>
   [@trevor-scheer](https://github.com/trevor-scheer) in [#1154](https://github.com/apollographql/graphql-tools/pull/1154)
+* Fix `delegateToSchema` to allow delegation to subscriptions with different root field names, allows
+  the use of the `RenameRootFields` transform with subscriptions,
+  pull request [#1104](https://github.com/apollographql/graphql-tools/pull/1104), fixes
+  [#997](https://github.com/apollographql/graphql-tools/issues/997).  <br/>
 
 ### 4.0.4